add AppliedCharge class
